requirements.txt
efficientnet_pytorch==0.7.1
keras==2.9.0
matplotlib==3.5.2
numpy==1.23.1
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
tensorflow==2.9.1
torch==1.9.0+cu111
torchvision==0.10.0+cu111
